pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "zarrio"
description = "A modern library for converting scientific data to Zarr format"
readme = "README.md"
authors = [
  {name = "Oceanum Developers", email = "developers@oceanum.science"}
]
maintainers = [
  {name = "Oceanum Developers", email = "developers@oceanum.science"}
]
classifiers=[
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "xarray>=0.18.0",
    "zarr>=2.10.0",
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "click>=8.0.0",
    "pyyaml>=5.4.0",
    "pyyaml>=5.4.0",
    "dask",
    "pydantic>2"
]
dynamic = ["version"]

[project.optional-dependencies]
oceanum = [
    "oceanum",
]
dev = [
    "pytest>=6.2.0",
    "pytest-cov>=2.12.0",
    "black>=21.0.0",
    "flake8>=3.9.0",
    "mypy>=0.910",
    "pre-commit>=2.13.0",
]
docs = [
    "sphinx>=4.0.0",
    "pydata-sphinx-theme"
]

[project.urls]
homepage = "https://github.com/oceanum/zarrio"
documentation = "https://github.com/oceanum/zarrio/blob/main/docs"
repository = "https://github.com/oceanum/zarrio"
issues = "https://github.com/oceanum/zarrio/issues"

[project.scripts]
zarrio = "zarrio.cli:main"

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml", "*.json"]

[tool.setuptools.dynamic]
version = {attr = "zarrio.__version__"}

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
